Fixed Gen3

+++ Gen3.java
@@ -1,8 +1,8 @@
 public class Gen3 {
     public static void main(String[] args) {
-        int min = Integer.parseInt(args[0]);
-		int max = Integer.parseInt(args[1]);
-		int range = max - min;
+        double min = double.parseDouble(args[0]);
+		double max = double.parseDouble(args[1]);
+		double range = max - min;
 		int rand1 = (int)(Math.random() * range) + min;
 		System.out.println(rand1);
 		int rand2 = (int)(Math.random() * range) + min;
